+++ backend/schemas/shared_schema.py
@@ -1,0 +1,3 @@
+from flask_marshmallow import Marshmallow
+
+ma = Marshmallow()
